Moved `FlowExt`
